requirements.txt
deepspeed==0.13.1
hydra-core==1.3.2
lightning==2.2.0
lightning-utilities==0.10.1
matplotlib==3.7.4
numpy==1.24.4
omegaconf==2.3.0
optuna==3.5.0
pandas==1.5.0
pillow==10.2.0
pytorch-lightning==2.2.0
scikit-learn==1.3.2
scipy==1.10.1
timm==0.9.12
torch==2.2.0
torchmetrics==1.3.0.post0
torchvision==0.17.0
tqdm==4.66.1
wandb==0.16.3